update api cart

+++ src/main/java/com/kelompok1/dailyyou/service/CartService.java
@@ -1,12 +1,13 @@
 package com.kelompok1.dailyyou.service;
 
 
+import com.kelompok1.dailyyou.configuration.exception.CartItemNotExistException;
 import com.kelompok1.dailyyou.model.dto.AddToCartDto;
 import com.kelompok1.dailyyou.model.dto.CartItemDto;
 import com.kelompok1.dailyyou.model.entity.Cart;
 import com.kelompok1.dailyyou.model.dto.CartDto;
 import com.kelompok1.dailyyou.model.entity.Product;
-//import com.kelompok1.dailyyou.model.entity.User;
+import com.kelompok1.dailyyou.model.entity.User;
 import com.kelompok1.dailyyou.repository.CartRepository;
 import com.kelompok1.dailyyou.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,8 +24,8 @@ import java.util.List;
 public class CartService {
     @Autowired
     private  CartRepository cartRepository;
-//
-//    private User user;
+
+    private User user;
 
     public CartService(){}
 
@@ -32,8 +33,8 @@ public class CartService {
         this.cartRepository = cartRepository;
     }
 
-    public void addToCart(AddToCartDto addToCartDto, Product product){
-        Cart cart = new Cart(product, addToCartDto.getProductQuantity());
+    public void addToCart(AddToCartDto addToCartDto, Product product, User user){
+        Cart cart = new Cart(product, addToCartDto.getProductQuantity(),user);
         cartRepository.save(cart);
     }
 
@@ -58,22 +59,24 @@ public class CartService {
     }
 
 
-    public void updateCartItem(AddToCartDto cartDto,Product product){
+    public void updateCartItem(AddToCartDto cartDto,User user,Product product){
         Cart cart = cartRepository.getOne(cartDto.getId());
         cart.setProductQuantity(cartDto.getProductQuantity());
         cart.setCreatedDate(new Date());
         cartRepository.save(cart);
     }
-    public void deleteCartItem(int id) {
+    public void deleteCartItem(int id,int userId) throws CartItemNotExistException {
+        if (!cartRepository.existsById(id))
+            throw new CartItemNotExistException("Cart id is invalid : " + id);
         cartRepository.deleteById(id);
 
     }
 
-    public void deleteCartItems() {
+    public void deleteCartItems(int userId) {
         cartRepository.deleteAll();
     }
 
-//    public void deleteUserCartItems(User user) {
-//        cartRepository.deleteByUser(user);
-//    }
+    public void deleteUserCartItems(User user) {
+        cartRepository.deleteByUser(user);
+    }
 }
